+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Features:
 Other changes:
 
 - *Backwards-incompatible*: ``validators.Length`` is no longer a subclass of ``validators.Range`` (:issue:`458`). Thanks :user:`deckar01` for the catch and patch.
+- Drop support for Python<=2.6.
 
 Deprecation/Removals: